+++ lib/modular_api.dart
@@ -11,6 +11,7 @@ export 'package:shelf/shelf.dart' show Middleware, Handler, Request, Response;
 // Core
 export 'src/core/modular_api.dart' show ModularApi, ModuleBuilder;
 export 'src/core/usecase/usecase.dart' show UseCase, Input, Output;
+export 'src/core/usecase/use_case_exception.dart' show UseCaseException;
 export 'src/core/usecase/usecase_test_handler.dart' show useCaseTestHandler;
 export 'src/core/oauth_handler.dart' show createOAuthTokenHandler;